Table cell classname -> style config + appearances
